+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 walkers 7.1.0
 ================
 - fix phantoms flying trait
+- add config option to overwrite the API Level
 
 walkers 7.0
 ================